Revert "自動ツイートのデータ集計機能を追加"
