+++ .env.ci
@@ -45,7 +45,8 @@ ADMIN_PASSWORD=
 # 記事投稿時のツイート用Twitter認証情報
 TWITTER_CONSUMER_KEY=
 TWITTER_CONSUMER_SECRET=
-TWITTER_BEARER_TOKEN=
+TWITTER_ACCESS_TOKEN=
+TWITTER_ACCESS_TOKEN_SECRET=
 
 # バックアップ用dropbox認証情報
 DROPBOX_AUTHORIZATION_TOKEN=
