+++ app/Http/Resources/Api/Mypage/Article.php
@@ -27,14 +27,8 @@ class Article extends JsonResource
             'created_at' => $this->created_at,
             'updated_at' => $this->updated_at,
             'url' => route('articles.show', $this->slug),
-            'metrics' => [
-                'totalViewCount' => $this->totalViewCount->count ?? 0,
-                'totalConversionCount' => $this->totalConversionCount->count ?? 0,
-                'totalRetweetCount' => $this->tweetLogSummary?->total_retweet_count ?? 0,
-                'totalReplyCount' => $this->tweetLogSummary?->total_reply_count ?? 0,
-                'totalLikeCount' => $this->tweetLogSummary?->total_like_count ?? 0,
-                'totalQuoteCount' => $this->tweetLogSummary?->total_quote_count ?? 0,
-            ],
+            'views' => $this->totalViewCount->count ?? 0,
+            'conversions' => $this->totalConversionCount->count ?? 0,
         ];
     }
 }
