+++ app/Models/ConversionCount.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace App\Models\Article;
+namespace App\Models;
 
-use App\Models\Article;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
-class ViewCount extends Model
+class ConversionCount extends Model
 {
     public const TYPE_DAILY = 1;
     public const TYPE_MONTHLY = 2;
@@ -22,10 +21,11 @@ class ViewCount extends Model
 
     public $timestamps = false;
 
-    protected $casts = [
-        'count' => 'integer',
-    ];
-
+    /*
+    |--------------------------------------------------------------------------
+    | リレーション
+    |--------------------------------------------------------------------------
+    */
     public function article(): BelongsTo
     {
         return $this->belongsTo(Article::class);
