+++ app/Models/ViewCount.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace App\Models\Article;
+namespace App\Models;
 
-use App\Models\Article;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
-class ConversionCount extends Model
+class ViewCount extends Model
 {
     public const TYPE_DAILY = 1;
     public const TYPE_MONTHLY = 2;
@@ -22,11 +21,10 @@ class ConversionCount extends Model
 
     public $timestamps = false;
 
-    /*
-    |--------------------------------------------------------------------------
-    | リレーション
-    |--------------------------------------------------------------------------
-    */
+    protected $casts = [
+        'count' => 'integer',
+    ];
+
     public function article(): BelongsTo
     {
         return $this->belongsTo(Article::class);
