+++ app/Repositories/ArticleRepository.php
@@ -18,7 +18,7 @@ use Illuminate\Support\LazyCollection;
 class ArticleRepository extends BaseRepository
 {
     private const COLUMNS = ['id', 'user_id', 'slug', 'title', 'post_type', 'contents', 'status', 'updated_at', 'created_at'];
-    private const RELATIONS = ['user', 'attachments', 'categories', 'tags', 'tweetLogSummary', 'totalViewCount', 'totalConversionCount'];
+    private const RELATIONS = ['user:id,name', 'attachments:id,attachmentable_id,attachmentable_type,path', 'categories:id,type,slug'];
     private const ORDER = ['updated_at', 'desc'];
 
     /**
@@ -420,10 +420,5 @@ class ArticleRepository extends BaseRepository
         );
 
         return $q->paginate($limit);
-    }
-
-    public function findByTitles(array $titles): Collection
-    {
-        return $this->model->active()->whereIn('title', $titles)->get();
     }
 }
