--- app/Services/TwitterAnalytics/InvalidTweetDataException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Services\TwitterAnalytics;
-
-class InvalidTweetDataException extends \Exception
-{
-}
