+++ config/twitter.php
@@ -3,5 +3,6 @@
 return [
     'consumer_key' => env('TWITTER_CONSUMER_KEY'),
     'consumer_secret' => env('TWITTER_CONSUMER_SECRET'),
-    'bearer_token' => env('TWITTER_BEARER_TOKEN'),
+    'access_token' => env('TWITTER_ACCESS_TOKEN'),
+    'access_token_secret' => env('TWITTER_ACCESS_TOKEN_SECRET'),
 ];
